Change the CentOS 7 repo to vault.centos.org
The CentOS Linux 7 reached the end of life (EOL) on June 30, 2024. For
the sandbox based on CentOS 7, we cannot get any update using
mirrors.centos.org. So change the mirror to vault.centos.org

Now it can work well on x86_64 cpu, but cannot work on ARM and other
CPUs.